Layout html skeleton and begin css styling
Filled in the html with the necessary divs and other elements to replicate
the example image, arranging and ordering everything in a sensible way.
Started some initial CSS styling, got the header well along, as well as
the button styling, plus I remembered to put in the box-sizing border-box
global change. Also saved the example image and style sheet to the repo.